# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AdWallet Africa API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kora/deposits/initialize:
    post:
      operationId: initializeKoraDeposit
      tags: [kora]
      summary: Initialize a mobile money or card deposit charge via Kora
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoraDepositInput"
      responses:
        "201":
          description: Deposit initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoraDepositResponse"
  /kora/deposits/{id}/status:
    get:
      operationId: getKoraDepositStatus
      tags: [kora]
      summary: Get status of a deposit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit status details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoraDepositDetails"
  /kora/deposits:
    get:
      operationId: listKoraDeposits
      tags: [kora]
      summary: List all user deposits
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoraDepositsListResponse"
  /kora/webhooks:
    post:
      operationId: processKoraWebhook
      tags: [kora]
      summary: Receive webhook updates from Kora
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoraWebhookInput"
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
...
components:
  schemas:
...
    MessageResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    KoraDepositInput:
      type: object
      required: [amount, currency, paymentMethod]
      properties:
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          enum: [GHS, NGN, KES]
        paymentMethod:
          type: string
          enum: [mobile_money, bank_transfer, card]
        provider:
          type: string
        phone:
          type: string
        email:
          type: string
    KoraDepositResponse:
      type: object
      required: [id, reference, status, amount, currency]
      properties:
        id:
          type: integer
        reference:
          type: string
        authorizationUrl:
          type: ["string", "null"]
        status:
          type: string
        amount:
          type: number
        currency:
          type: string
    KoraDepositDetails:
      type: object
      required: [id, reference, status, amountLocal, localCurrency, fee, adwalletFee, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        reference:
          type: string
        status:
          type: string
        amountLocal:
          type: number
        localCurrency:
          type: string
        amountUsd:
          type: ["number", "null"]
        exchangeRate:
          type: ["number", "null"]
        fee:
          type: number
        adwalletFee:
          type: number
        paymentMethod:
          type: string
        provider:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    KoraDepositsListResponse:
      type: object
      required: [deposits, total, page]
      properties:
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/KoraDepositDetails"
        total:
          type: integer
        page:
          type: integer
...
    KoraWebhookInput:
      type: object
      additionalProperties: true