# lib/api-zod/src/generated/types/koraDepositsListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KoraDepositDetails } from './koraDepositDetails';

export interface KoraDepositsListResponse {
  deposits: KoraDepositDetails[];
  total: number;
  page: number;
}