# lib/api-zod/src/generated/types/koraWebhookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KoraWebhookInput { [key: string]: unknown }